tests: reorganize tests hierarchically
