+++ tests/.gitignore
@@ -16,26 +16,26 @@ TAGS
 *.xml
 .arch-inventory
 .arch-ids
-aspathtest
-ecommtest
-heavy
-heavythread
-heavywq
-tabletest
-test-timer-correctness
-test-timer-performance
-testbgpcap
-testbgpmpath
-testbgpmpattr
-testbuffer
-testchecksum
-testcli
-testmemory
-testprivs
-testsegv
-testsig
-teststream
-testnexthopiter
-testcommands
-test-commands-defun.c
-site.exp
+/bgpd/test_aspath
+/bgpd/test_capability
+/bgpd/test_ecommunity
+/bgpd/test_mp_attr
+/bgpd/test_mpath
+/lib/cli/test_cli
+/lib/cli/test_commands
+/lib/cli/test_commands_defun.c
+/lib/test_buffer
+/lib/test_checksum
+/lib/test_heavy
+/lib/test_heavy_thread
+/lib/test_heavy_wq
+/lib/test_memory
+/lib/test_nexthop_iter
+/lib/test_privs
+/lib/test_srcdest_table
+/lib/test_segv
+/lib/test_sig
+/lib/test_stream
+/lib/test_table
+/lib/test_timer_correctness
+/lib/test_timer_performance
