+++ tests/bgpd/test_capability.c
