+++ tests/helpers/c/prng.c
