+++ tests/lib/cli/common_cli.h
