+++ tests/lib/cli/test_cli.in
