+++ tests/lib/cli/test_cli.refout
