+++ tests/lib/cli/test_commands.c
