+++ tests/lib/cli/test_commands.in
