+++ tests/lib/cli/test_commands.refout
