+++ tests/lib/test_buffer.c
