+++ tests/lib/test_checksum.c
