+++ tests/lib/test_memory.c
