+++ tests/lib/test_nexthop_iter.c
