+++ tests/lib/test_privs.c
