+++ tests/lib/test_sig.c
