+++ tests/lib/test_stream.c
