+++ tests/lib/test_table.c
