+++ tests/lib/test_timer_correctness.c
