+++ tests/lib/test_timer_performance.c
